Más avances con el servicio 5, ya casi está, falta el 14

+++ src/main/java/domain/models/entities/services/calculadorasGradoDeConfianza/gradoCalculadorEquipo5/entities/ComunidadDevuelta.java
@@ -8,5 +8,5 @@ import lombok.Setter;
 public class ComunidadDevuelta {
     public ComunidadApi5 comunidad;
     public double nuevoPuntaje;
-    public double gradoDeConfianzaActual;
+    public Integer gradoDeConfianzaActual;
 }
